Refactor kolibri.tests.test_cli to new utils test package
