+++ test/bdd/bdd_basic.py
@@ -3,7 +3,7 @@ BDD tests
 """
 from __future__ import absolute_import, print_function, unicode_literals
 
-from test.test_cli import KolibriTestBase
+from kolibri.utils.tests.base import KolibriTestBase
 
 import logging
